refactor: remove integration tests

+++ src/domain/usecases/impl/delete-card-usecase.ts
@@ -9,9 +9,20 @@ export class DeleteCard implements DeleteCardUseCase {
   constructor(private readonly cardRepository: CardRepository) {}
 
   async execute(input: DeleteCardInput): Promise<DeleteCardOutput> {
+    const card = await this.cardRepository.findById(input.id);
+
+    if (!card) {
+      return {
+        exists: false,
+        success: false,
+      };
+    }
+
     const result = await this.cardRepository.delete(input.id);
+
     return {
       success: result,
+      exists: true,
     };
   }
 }
